Move scheduler logic to main folder
