+++ interface/urls.py
@@ -22,5 +22,4 @@ from django.conf.urls.static import static
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include("main.urls")),
-    path('scheduler/', include("scheduler.urls")),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
