+++ main/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from .models import CrawlRequest, CrawlerInstance, CrawlerQueue, CrawlerQueueItem
+from .models import CrawlRequest, CrawlerInstance, CrawlerQueue, CrawlerQueueItem, Task
 
 
 class CrawlerInstanceSerializer(serializers.ModelSerializer):
@@ -47,3 +47,13 @@ class CrawlerQueueSerializer(serializers.ModelSerializer):
         model = CrawlerQueue
 
         fields = '__all__'
+
+class TaskSerializer(serializers.ModelSerializer):
+    crawler_name = serializers.ReadOnlyField(source='crawl_request.source_name')
+
+    class Meta:
+        model = Task
+        read_only_fields = ['id', 'creation_date', 'last_modified']
+
+        fields = ['id', 'creation_date', 'last_modified', 'crawl_request', 'runtime','crawler_name', 
+                'crawler_queue_behavior', 'repeat_mode', 'personalized_repetition_mode']
