+++ main/staticfiles/js/scheduler/services.js
@@ -6,7 +6,7 @@ services.save_new_scheduling = function (new_scheduling_config) {
     let parsed_data = JSON.stringify(new_scheduling_config);
 
     $.ajax({
-        url: '/scheduler/tasks/',
+        url: '/api/tasks/',
         type: 'post',
         contentType: "application/json; charset=utf-8",
         dataType: "json",
@@ -23,13 +23,13 @@ services.save_new_scheduling = function (new_scheduling_config) {
 }
 
 services.get_tasks_in_interval = function (start_date, end_date) {
-    let tasks_by_date;
     $.ajax({
-        url: `/scheduler/tasks/filter?start_date=${start_date}&end_date=${end_date}`,
+        url: `/api/tasks/filter?start_date=${start_date}&end_date=${end_date}`,
         type: 'get',
         async: false,
         success: function (data) {
             tasks_by_date = data;
+            
         },
         error: function (data) {
             console.error(data.responseText);
@@ -42,7 +42,7 @@ services.get_tasks_in_interval = function (start_date, end_date) {
 services.get_task = function (task_id) {
     let task;
     $.ajax({
-        url: `/scheduler/tasks/${task_id}`,
+        url: `/api/tasks/${task_id}`,
         type: 'get',
         async: false,
         success: function (data) {
@@ -57,6 +57,7 @@ services.get_task = function (task_id) {
 }
 
 services.update_tasks = function (tarks_ids) {
+
     tasks = {};
     let task_id;
     for (let i=0;i<tarks_ids.length;i++) {
